feat(www): moves faq page to docs
