+++ www/docs/main/faq.mdx
@@ -1,0 +1,12 @@
+---
+id: faq
+title: Frequently Asked Questions
+sidebar_label: Frequently Asked Questions
+slug: /faq
+---
+
+import { Faq } from '../../src/components/Faq.tsx';
+
+A collection of the most commonly asked questions in regards to everything tRPC! If you have a question that isn't answered here, you can join our [Discord](https://discord.com/invite/wzaMgEJkcf) and we'll be happy to help!
+
+<Faq />
